@@ -2,10 +2,24 @@
   "name": "mongoose-collection-naming-plugin",
   "version": "1.1.5",
   "description": "Mongoose plugin to name normalize collections names based on naming convention camelCase, snake_case, PascalCase, ...etc",
-  "main": "dist/index.js",
+  "type": "module",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
   "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "require": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.cjs"
+      }
+    }
+  },
   "scripts": {
-    "build": "rimraf dist && tsc -p tsconfig.build.json",
+    "build": "rimraf dist && rollup -c",
     "lint": "eslint {lib,test}/**/*.ts",
     "lint:fix": "eslint {lib,test}/**/*.ts --fix",
     "format": "prettier --write {lib,test}/**/*.ts",
@@ -30,32 +44,37 @@
   "devDependencies": {
     "@commitlint/cli": "^19.6.1",
     "@commitlint/config-conventional": "^19.6.0",
-    "@stylistic/eslint-plugin-js": "^2.12.1",
-    "@stylistic/eslint-plugin-ts": "^2.12.1",
+    "@rollup/plugin-commonjs": "^29.0.0",
+    "@rollup/plugin-node-resolve": "^16.0.3",
+    "@rollup/plugin-typescript": "^12.3.0",
+    "@stylistic/eslint-plugin": "^5.5.0",
     "@types/node": "^22.10.3",
     "@types/pluralize": "^0.0.33",
-    "@typescript-eslint/eslint-plugin": "8.19.0",
-    "@typescript-eslint/parser": "8.19.0",
-    "@vitejs/plugin-vue": "^5.2.1",
+    "@typescript-eslint/eslint-plugin": "8.46.3",
+    "@typescript-eslint/parser": "8.46.3",
     "@vitest/coverage-istanbul": "^2.1.8",
     "conventional-changelog-atom": "^5.0.0",
-    "eslint": "^8.57.1",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "^3.7.0",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-n": "^17.15.1",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint": "^9.39.1",
+    "eslint-config-prettier": "10.1.8",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-n": "^17.23.1",
+    "eslint-plugin-prettier": "^5.5.4",
     "eslint-plugin-security": "^3.0.1",
-    "eslint-plugin-sonarjs": "^3.0.1",
+    "eslint-plugin-sonarjs": "^3.0.5",
+    "globals": "^16.5.0",
     "husky": "^9.1.7",
     "lint-staged": "15.3.0",
     "mongodb-memory-server": "^10.1.3",
     "mongoose": "^8.9.3",
-    "prettier": "3.4.2",
-    "rimraf": "6.0.1",
+    "prettier": "3.6.2",
+    "rimraf": "6.1.0",
+    "rollup": "^4.53.2",
+    "rollup-plugin-dts": "^6.2.3",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "typescript": "5.7.2",
+    "typescript-eslint": "^8.46.3",
     "vitest": "^2.1.8"
   },
   "peerDependencies": {
